+++ docs/release-notes.rst
@@ -20,9 +20,14 @@
 Azure Client Release Notes
 ==========================
 
+Version 1.2.21, 20 September 2018
+---------------------------------
+* Upgrade Spring-Boot version (ACUMOS-1754)
+* Upgrade Springfox version
+
 Version 1.2.20, 17 September 2018
 ---------------------------------
-Azure-client logs not written to location same as others(ACUMOS-1635)
+* Azure-client logs not written to location same as others (ACUMOS-1635)
 
 Version 1.2.18, 12 September 2018
 ---------------------------------